Refactor LogsTableRow into its own folder
