+++ studio/src/pages/RequestorPage/LogsTable/LogsTableRow/LogsTableRow.tsx
@@ -1,0 +1,14 @@
+import { type LogEntry, isNeonEvent } from "../types";
+import { LogRow } from "./LogRow";
+import { NeonEventRow } from "./NeonEventRow";
+
+type LogRowProps = {
+  log: LogEntry;
+};
+
+export function LogsTableRow({ log }: LogRowProps) {
+  if (isNeonEvent(log)) {
+    return <NeonEventRow log={log} />;
+  }
+  return <LogRow log={log} />;
+}
