+++ studio/src/pages/RequestorPage/LogsTable/LogsTableRow/index.ts
@@ -1,0 +1,1 @@
+export { LogsTableRow } from "./LogsTableRow";
