+++ studio/src/pages/RequestorPage/LogsTable/LogsTableRow/shared.ts
@@ -1,0 +1,7 @@
+export function formatTimestamp(timestamp: Date) {
+  return timestamp.toLocaleTimeString([], {
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+  });
+}
